api/requirements.txt
torchvision==0.9.0
torch==1.8.0
gunicorn==20.0.4
requests==2.24.0
Flask==1.1.2
Pillow==8.1.2
